kiril 3 lab final

+++ kirusha_proxy/3lab/creature.cpp
@@ -20,17 +20,17 @@ Creature::Creature(Swamp* s, c_type t) : _type(t), _swamp(s)
 	{
 		_name = "рыба-селедка";	//был микроб но тогда его никто не ест
 		_type_to_evolve = c_type::REPT;
-		_max_age = 6;
+		_max_age = 30;
 		_type_to_kill.push_back(c_type::MICRO);
 		_w_breed = 80;
-		_w_kill = 40;
-		_w_evolve = 30;
+		_w_kill = 20;
+		_w_evolve = 20;
 		break;
 	}
 	case REPT:
 	{
 		_name = "рептилияяя";
-		_max_age = 25;
+		_max_age = 40;
 		_type_to_evolve = c_type::TEPL;
 		_type_to_kill.push_back(c_type::TEPL);
 		_type_to_kill.push_back(c_type::REPT);
@@ -39,7 +39,7 @@ Creature::Creature(Swamp* s, c_type t) : _type(t), _swamp(s)
 		_type_to_kill.push_back(c_type::MICRO);
 		_type_to_kill.push_back(c_type::MICRO);
 		_w_breed = 70;
-		_w_kill = 30;
+		_w_kill = 15;
 		_w_evolve = 10;
 		break;
 	}
@@ -59,7 +59,7 @@ Creature::Creature(Swamp* s, c_type t) : _type(t), _swamp(s)
 		_type_to_kill.push_back(c_type::MICRO);
 		_type_to_kill.push_back(c_type::HUM);
 		_w_breed = 60;
-		_w_kill = 40;
+		_w_kill = 10;
 		_w_evolve = 10;
 		break;
 	}
@@ -76,7 +76,7 @@ Creature::Creature(Swamp* s, c_type t) : _type(t), _swamp(s)
 		_type_to_kill.push_back(c_type::TEPL);
 		_type_to_kill.push_back(c_type::HUM);
 		_w_breed = 30;
-		_w_kill = 50;
+		_w_kill = 12;
 		_w_evolve = 15;
 		break;
 	}
@@ -94,7 +94,7 @@ Creature::Creature(Swamp* s, c_type t) : _type(t), _swamp(s)
 		_type_to_kill.push_back(c_type::MICRO);
 		_type_to_kill.push_back(c_type::MICRO);
 		_w_breed = 30;
-		_w_kill = 60;
+		_w_kill = 15;
 		_w_evolve = 20;
 		break;
 	}
@@ -107,7 +107,7 @@ Creature::Creature(Swamp* s, c_type t) : _type(t), _swamp(s)
 		_type_to_kill.push_back(c_type::TEPL);
 		_type_to_kill.push_back(c_type::MICRO);
 		_w_breed = 10;
-		_w_kill = 20;
+		_w_kill = 5;
 		_w_evolve = 5;
 		break;
 	}
@@ -122,7 +122,8 @@ Creature::Creature(Swamp* s, c_type t) : _type(t), _swamp(s)
 		_type_to_kill.push_back(c_type::HUM);
 		_type_to_kill.push_back(c_type::HUMPAZ);
 		_type_to_kill.push_back(c_type::SUPRAZ);
-		_w_kill = 90;
+		_w_kill = 100;
+		_w_breed = 20;
 		break;
 	}
 
@@ -138,7 +139,9 @@ Creature::Creature(Swamp* s, c_type t) : _type(t), _swamp(s)
 
 bool Creature::breed()
 {
+	//std::cout << "pre breed" << std::endl;
 	std::lock_guard<std::mutex> lock(_inside);
+	//std::cout << "in breed" << std::endl;
 	if (_alive) {
 		if (_type == c_type::L)
 			return true;
@@ -154,8 +157,9 @@ bool Creature::breed()
 
 bool Creature::grow_old()
 {
+	//std::cout << "pre grow" << std::endl;
 	std::lock_guard<std::mutex> lock(_inside);
-	std::cout << "in grow" << std::endl;
+	//std::cout << "in grow" << std::endl;
 	if (_alive) {
 		_age += 1;
 		switch (_type) {
@@ -182,24 +186,28 @@ bool Creature::grow_old()
 bool Creature::kill() const
 {
 	if (_alive) {
-		int id = rand() % _type_to_kill.size();
-		c_type type_to_kill = _type_to_kill.at(id);
-		Creature* c = _swamp->get_creature(type_to_kill);
-		if (c != nullptr) {
-			if (c->die())
-				std::cout << _name << " killed " << c->_name << std::endl;
-			else
-				std::cout << _name << " tried to kill " << c->_name << " but it's immortal" << std::endl;
-			c->unlock();
+		if (_type_to_kill.size() > 0) {
+			int id = rand() % _type_to_kill.size();
+			c_type type_to_kill = _type_to_kill.at(id);
+			Creature* c = _swamp->get_creature(type_to_kill);
+			if (c != nullptr) {
+				if (c->die())
+					std::cout << _name << " killed " << c->_name << std::endl;
+				else
+					std::cout << _name << " tried to kill " << c->_name << " but it's immortal" << std::endl;
+				c->unlock();
+			}
+			std::cout << _name << " wanna kill someone but they're already dead :((((" << std::endl;
 		}
-		std::cout << _name << " wanna kill someone but they're already dead :((((" << std::endl;
 	}
 	return false;
 }
 
 bool Creature::evolve()
 {
+	//std::cout << "pre evolve" << std::endl;
 	std::lock_guard<std::mutex> lock(_inside);
+	//std::cout << "in evolve" << std::endl;
 	if (_alive && _type != c_type::ESS) {
 		Creature* c = new Creature(_swamp, _type_to_evolve);
 		_age += 5;
@@ -213,13 +221,14 @@ bool Creature::evolve()
 
 bool Creature::die()
 {
+	//std::cout << "pre die" << std::endl;
 	std::lock_guard<std::mutex> lock(_inside);
+	//std::cout << "in die" << std::endl;
 	if (_type == c_type::ESS)
 		return false;
 	_alive = false;
 	//std::cout << _name << " died :(" << std::endl;
 	return true;
 }
-
